+++ tests/TestCase.php
@@ -2,8 +2,10 @@
 
 namespace Tests;
 
+use Illuminate\Support\Facades\Route;
 use Laravel\FakeId\ServiceProvider;
 use Orchestra\Testbench\TestCase as BaseTestCase;
+use Tests\Stubs\Post;
 
 /**
  * Class TestCase
@@ -25,17 +27,5 @@ abstract class TestCase extends BaseTestCase
         return [
             ServiceProvider::class,
         ];
-    }
-
-    /**
-     * Define environment setup.
-     *
-     * @param  \Illuminate\Foundation\Application $app
-     *
-     * @return void
-     */
-    protected function getEnvironmentSetUp($app)
-    {
-        // Define your environment setup.
     }
 }