+++ CHANGELOG.md
@@ -26,9 +26,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+
+## [1.0.0] - 2019-11-11
+
+### Added
+- Add interface `Laravel\FakeId\Contracts\Manager`
+- Add driver `Laravel\FakeId\Drivers\PrefixDriver`
+- Add more test scripts
+- Add feature allowed specific driver by class name in config file
+
+### Changed
+- Rename interface `Laravel\FakeId\Drivers\DriverInterface` to `Laravel\FakeId\Contracts\Driver`
+- Change coding style to PSR-12
+
+
+
+
+
 ## 0.1.0 - 2018-10-08
 
 First release
 
 
-[Unreleased]: https://github.com/oanhnn/laravel-fakeid/compare/v0.1.0...develop
+[Unreleased]: https://github.com/oanhnn/laravel-fakeid/compare/v1.0.0...develop
+[1.0.0]:      https://github.com/oanhnn/laravel-fakeid/compare/v0.1.0...v1.0.0